requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==2.3.7
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.3
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
idna==3.4
urllib3==2.0.4
blinker==1.6.2 